Add investigate route with SSE streaming
