+++ apps/api/index.ts
@@ -1,13 +1,17 @@
-import "dotenv/config";
-import { readdir, readFile } from "node:fs/promises";
+import { config } from "dotenv";
+config({ path: new URL("../../.env", import.meta.url).pathname });
+import { mkdir, readdir, readFile, writeFile } from "node:fs/promises";
 import { dirname, join } from "node:path";
 import { fileURLToPath } from "node:url";
 import { serve } from "@hono/node-server";
 import { Hono } from "hono";
 import { cors } from "hono/cors";
+import { streamSSE } from "hono/streaming";
+import { investigate } from "@mev/agent";
 import { Outcome, RootCause, TradeVerdict } from "@mev/shared";
 import type {
     GetTradeOutput,
+    SSEEvent,
     TradeListItem,
     TradeReport,
 } from "@mev/shared";
@@ -100,6 +104,32 @@ app.get("/trades/:tx_hash", async (c) => {
     const report = await readReport(tx_hash);
     if (!report) return c.json({ error: "Report not found" }, 404);
     return c.json(report);
+});
+
+app.post("/trades/:tx_hash/investigate", async (c) => {
+    const tx_hash = c.req.param("tx_hash");
+
+    return streamSSE(c, async (stream) => {
+        const onEvent = async (event: SSEEvent) => {
+            await stream.writeSSE({ data: JSON.stringify(event) });
+        };
+
+        try {
+            const report = await investigate(tx_hash, onEvent);
+            if (report) {
+                await mkdir(REPORTS_DIR, { recursive: true });
+                await writeFile(
+                    join(REPORTS_DIR, `${tx_hash}.json`),
+                    JSON.stringify(report, null, 2)
+                );
+            }
+        } catch (err) {
+            const message = err instanceof Error ? err.message : String(err);
+            await stream.writeSSE({
+                data: JSON.stringify({ type: "error", message } satisfies SSEEvent),
+            });
+        }
+    });
 });
 
 const PORT = 3001;
